ml_service/requirements.txt
python-dotenv
requests
supabase
sentence-transformers
numpy
flask
pandas